Refactor metadata classes
To accommodate future development and support for other storage
models, the metadata core has been refactored into separate
storage and manager files.  In addition a set of custom errors
has been introduced that are not tied to the storage model.

The MetadataManager methods have changed for easier use and the
MetadataStore methods and FileMetadataStore implementation have
been redefined to be completely specific to storage.

All refactored code has also added type hints in method signatures
and all tests have  been updated and rewritten.
